% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{This is the number of the state and it must be a numerical value
or it will end up to an error}

\item{year}{this is the year when the accidents happened and it must be numerical
value}
}
\value{
it returns a graphical representation for the accidents in the \code{state.num}
during \code{year}
}
\description{
This function gives a representation on a map for the accidents on a specific state
and year
}
\note{
You should install dplyr , graghics and maps packages
}
\examples{
\dontrun{
map <- fars_map_state(state.num = 14 ,year = 2015)
}

}
