% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{it is the path for the csv file or the data you want to read through}
}
\value{
a tibble data frame that contains the data from \code{filename}
}
\description{
this function is for reading the a data frame or csv file and return a tibble data frame
}
\examples{
\dontrun{
df <- fars_read("accident_2015.csv")
df <- fars_read( filename = "accident_2015.csv")
}

}
