% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize specific years data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{represents a vector or a list containing years which should data be
extracted from}
}
\value{
a tibble that contains the number of accidents per month and year for
several years which are provided through \code{years}
}
\description{
This function summarizes the data of the year and provide the number of accidents
}
\examples{
\dontrun{
accident_num <- fars_summarize_years(2014)
}

}
