% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{file name with a specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{it refers to a year which you want to creat file name with.
it has to be a number or a string that contain a numer or it will end up with
a waring and NA will be the return value from the function .}
}
\value{
it will return a string that represents a file name
}
\description{
This function creates a file name for a particular year .
}
\examples{
\dontrun{
file_name <- make_filename(2015)
file_name <- make_filename("2015")
file_name <- make_filename(year = "2015")
file_name <- make_filename(year = 2015)
}

}
